refactor(setup): extract the Pi bridge template into pi_bridge module
The module-size ratchet gate failed on the previous commit: setup.py
reached 5268 lines against its grandfathered 5225 budget (+43). The
gate's contract is shrink-only, so the right move is decomposition
rather than a budget bump.

Move the pure TypeScript rendering into the new
ouroboros.cli.commands.pi_bridge module as
pi_ooo_bridge_source_text(command, args). setup.py keeps
_detect_pi_bridge_dispatch_entry and _pi_ooo_bridge_source_text as a
thin wrapper, so callers and tests keep the same API. The rendered
bridge is byte-identical before and after the extraction, and setup.py
drops to 5127 lines, back under budget.
