+++ src/ouroboros/cli/commands/setup.py
@@ -46,6 +46,7 @@ from ouroboros.cli.commands.claude_setup import (
 from ouroboros.cli.commands.claude_setup import (
     setup_claude_sdk as _setup_claude_sdk,
 )
+from ouroboros.cli.commands.pi_bridge import pi_ooo_bridge_source_text
 from ouroboros.cli.commands.setup_atomic_restore import restore_hermes, restore_hermes_receipt
 from ouroboros.cli.formatters import console
 from ouroboros.cli.formatters.panels import (
@@ -3604,148 +3605,7 @@ def _detect_pi_bridge_dispatch_entry() -> tuple[str, list[str]]:
 def _pi_ooo_bridge_source_text() -> str:
     """Return the managed Pi extension source for ``ooo`` frontdoor dispatch."""
     command, args = _detect_pi_bridge_dispatch_entry()
-    default_command = json.dumps(command)
-    default_args = json.dumps(args)
-    return f"""/**
- * Ouroboros ooo bridge for Pi.
- *
- * Managed by `ouroboros setup --runtime pi`.
- * Routes exact-prefix `ooo ...` inputs from interactive Pi into Ouroboros'
- * shared skill dispatcher instead of sending them to the model as chat.
- * The registered `/ooo` command also provides TAB argument completion for
- * dispatchable subcommands and Seed files.
- */
-import {{ readdirSync }} from "node:fs";
-import {{ homedir }} from "node:os";
-import * as path from "node:path";
-import type {{ ExtensionAPI, ExtensionContext }} from "@earendil-works/pi-coding-agent";
-
-type CompletionItem = {{ value: string; label: string; description?: string }};
-
-// Dispatchable `ooo` subcommands: the packaged skills whose SKILL.md declares
-// `mcp_tool` frontmatter, i.e. the commands the shared dispatcher can execute
-// as one deterministic MCP call. Kept in sync by a unit test.
-const DISPATCHABLE_COMMANDS: Array<{{ cmd: string; description: string }}> = [
-  {{ cmd: "auto", description: "Interview, Seed generation, and run handoff" }},
-  {{ cmd: "interview", description: "Socratic interview to clarify requirements" }},
-  {{ cmd: "run", description: "Execute a Seed specification" }},
-  {{ cmd: "seed", description: "Generate a Seed from an interview session" }},
-  {{ cmd: "status", description: "Session status and drift check" }},
-  {{ cmd: "ralph", description: "Evolutionary loop until QA passes" }},
-];
-
-function matchesPrefix(value: string, prefix: string): boolean {{
-  return value.toLowerCase().startsWith(prefix.toLowerCase());
-}}
-
-function seedFileCompletions(prefix: string): CompletionItem[] | null {{
-  const seedsDir = path.join(homedir(), ".ouroboros", "seeds");
-  let names: string[];
-  try {{
-    names = readdirSync(seedsDir);
-  }} catch {{
-    return null;
-  }}
-  // Completion values carry the absolute store path: the `run` dispatcher
-  // resolves relative seed paths against the Pi session cwd, so a bare
-  // filename would only execute when the same file also exists there.
-  // Names containing whitespace are skipped because the dispatcher
-  // tokenizes the dispatched command on whitespace.
-  const items = names
-    .filter((name) => name.endsWith(".yaml") || name.endsWith(".yml"))
-    .filter((name) => !/\\s/.test(name) && matchesPrefix(name, prefix))
-    .sort()
-    .map((name) => ({{
-      value: path.join(seedsDir, name),
-      label: name,
-      description: "Seed file",
-    }}));
-  return items.length > 0 ? items : null;
-}}
-
-// TAB completions for `/ooo <TAB>`: dispatchable subcommands for the first
-// argument, Seed files from `~/.ouroboros/seeds/` for `ooo run <TAB>`
-// (inserted as absolute paths so the value executes from any session cwd).
-function argumentCompletions(argumentPrefix: string): CompletionItem[] | null {{
-  const tokens = argumentPrefix.replace(/^\\s+/, "").split(/\\s+/);
-  const completing = tokens[tokens.length - 1] ?? "";
-  if (tokens.length <= 1) {{
-    const items = DISPATCHABLE_COMMANDS.filter((entry) =>
-      matchesPrefix(entry.cmd, completing),
-    ).map((entry) => ({{ value: entry.cmd, label: entry.cmd, description: entry.description }}));
-    return items.length > 0 ? items : null;
-  }}
-  if (tokens[0].toLowerCase() === "run" && tokens.length === 2) {{
-    return seedFileCompletions(completing);
-  }}
-  return null;
-}}
-
-const COMMAND_RE = /^\\s*ooo(?:\\s+|$)/i;
-const UNSUPPORTED_DISPATCH_EXIT_CODE = 78;
-const TIMEOUT_MS = Number(process.env.OUROBOROS_PI_BRIDGE_TIMEOUT_MS || 6 * 60 * 60 * 1000);
-const DEFAULT_COMMAND = {default_command};
-const DEFAULT_ARGS = {default_args};
-
-function ouroborosEntry(): {{ command: string; args: string[] }} {{
-  if (process.env.OUROBOROS_CLI) return {{ command: process.env.OUROBOROS_CLI, args: [] }};
-  return {{ command: DEFAULT_COMMAND, args: DEFAULT_ARGS }};
-}}
-
-function outputText(stdout: string, stderr: string): string {{
-  const out = stdout.trim();
-  const err = stderr.trim();
-  if (out && err) return `${{out}}\\n\\n${{err}}`;
-  return out || err || "(no output)";
-}}
-
-async function dispatch(pi: ExtensionAPI, text: string, ctx: ExtensionContext): Promise<boolean> {{
-  ctx.ui.notify(`Ouroboros dispatch: ${{text}}`, "info");
-  const entry = ouroborosEntry();
-  const result = await pi.exec(
-    entry.command,
-    [...entry.args, "dispatch", "--runtime", "pi", "--cwd", ctx.cwd, text],
-    {{ cwd: ctx.cwd, timeout: TIMEOUT_MS }},
-  );
-  if (result.code === UNSUPPORTED_DISPATCH_EXIT_CODE) {{
-    ctx.ui.notify(`Ouroboros did not claim command; continuing in Pi`, "info");
-    return false;
-  }}
-  const body = outputText(result.stdout || "", result.stderr || "");
-  pi.sendMessage({{
-    customType: "ouroboros",
-    content: body,
-    display: true,
-    details: {{ command: text, exitCode: result.code }},
-  }});
-  if (result.code !== 0) {{
-    ctx.ui.notify(`Ouroboros dispatch failed (${{result.code ?? "unknown"}})`, "error");
-  }}
-  return true;
-}}
-
-export default function ouroborosBridge(pi: ExtensionAPI) {{
-  pi.registerCommand("ooo", {{
-    description: "Dispatch an Ouroboros ooo command",
-    getArgumentCompletions: argumentCompletions,
-    handler: async (args, ctx) => {{
-      const text = `ooo ${{args}}`.trim();
-      await dispatch(pi, text, ctx);
-    }},
-  }});
-
-  pi.on("input", async (event, ctx) => {{
-    if (event.source === "extension") {{
-      return {{ action: "continue" }};
-    }}
-    if (!COMMAND_RE.test(event.text)) {{
-      return {{ action: "continue" }};
-    }}
-    const handled = await dispatch(pi, event.text.trim(), ctx);
-    return {{ action: handled ? "handled" : "continue" }};
-  }});
-}}
-"""
+    return pi_ooo_bridge_source_text(command=command, args=args)
 
 
 def _install_pi_ooo_bridge() -> bool:
